requirements.txt
# Manuscript specific packages
python-docx
pyyaml

# Code Specific Packages